+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `:vips` will be documented in this file.
 - logging with PSR-3 Logger Interface [Kleis Auke Wolthuizen]
 - switch to PSR2 formatting [Kleis Auke Wolthuizen]
 - add sig.php example [John Cupitt]
+- add Vips\Image::debugLogger() sample logger [John Cupitt]
 
 ### Deprecated
 - removed `\Enum` from enum names